+++ CHANGELOG.md
@@ -6,10 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SModalCard` component implementation.
 
 ## 2.0.0-beta.16 - 2023-10-11
 ### Added
-- `SModalCard` component implementation.
 - `SModal`, `SModalTitle` and `SModalDescription` components implementation.
 - `SBreadcrumbs` and `SBreadcrumbsItem` components implementation.
 - `SToast` component implementation.