+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.0.0-beta.18 - 2023-10-13
 ### Added
 - Style variations with `Class Variance Authority`.
 - Actions slot on `SCard` and `SModalCard`.